5th Semester/Affective Computing/Project/requirements.txt
PyYAML~=6.0.1
torch~=2.1.2
torchaudio~=2.1.2
scipy~=1.11.4
numpy~=1.26.2
black~=23.11.0
matplotlib~=3.8.2
pandas~=2.1.3
librosa~=0.10.1
resampy~=0.4.2
transformers~=4.35.2
tensorboardX~=2.6.2.2
scikit-learn~=1.3.2
sentencepiece~=0.1.99
tqdm~=4.66.1
joblib~=1.3.2
